➕ Support Drag and Select feature
